fixed bug and failing tests
